Move mixins/table-pager into addon
- Include reexports in /app till client apps update their imports
- Fixes #36
